requirements.txt
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-bulma==0.8.3
django-compressor==4.4
django-libsass==0.9
djangorestframework==3.14.0
libsass==0.22.0
pytz==2023.3
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
